requirements.txt
beautifulsoup4
selenium
lxml
pytest-playwright
polars
pyarrow
tqdm
python-dotenv